This is a multi linear regression used for the independent variables being the continous.
This is an example of well proof of the multi linear regression where the independent variables are independent. Here I have used the sklearn module to depict the linear regression. There is a clear definition is been provided about the over fitting and multi col-linearity issues about the individual variables.